+++ database/migrations/2015_07_30_081728_create_offices_table.php
@@ -12,21 +12,31 @@ class CreateOfficesTable extends Migration
      */
     public function up()
     {
-        Schema::create('offices', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->string('address');
-            $table->integer('status_id')->unsigned();
-            $table->timestamps();
+        if(Schema::hasTable('status') && Schema::hasTable('users') && Schema::hasTable('offices')){
+            Schema::table('offices', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned();
+                $table->string('address');
+                $table->integer('status_id')->unsigned();
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
+                $table->foreign('user_id')->references('id')->on('users')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
 
-            $table->foreign('status_id')->references('id')->on('status')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
-        });
+                $table->foreign('status_id')->references('id')->on('status')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
+            });
+        }else{
+            Schema::create('offices', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned();
+                $table->string('address');
+                $table->integer('status_id')->unsigned();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
