+++ database/migrations/2015_07_30_081801_create_appointments_table.php
@@ -12,27 +12,39 @@ class CreateAppointmentsTable extends Migration
      */
     public function up()
     {
-        Schema::create('appointments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('patient_id')->unsigned();
-            $table->integer('office_id')->unsigned();
-            $table->dateTime('appiontment_date');
-            $table->decimal('amount',10,2);
-            $table->integer('status_id')->unsigned();
-            $table->timestamps();
+         if(Schema::hasTable('status') && Schema::hasTable('patients') && Schema::hasTable('offices') && Schema::hasTable('appointments')){
+            Schema::table('appointments', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('patient_id')->unsigned();
+                $table->integer('office_id')->unsigned();
+                $table->dateTime('appiontment_date');
+                $table->decimal('amount',10,2);
+                $table->integer('status_id')->unsigned();
+                $table->timestamps();
 
-            $table->foreign('patient_id')->references('id')->on('patients')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
+                $table->foreign('patient_id')->references('id')->on('patients')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
 
-            $table->foreign('office_id')->references('id')->on('offices')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
+                $table->foreign('office_id')->references('id')->on('offices')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
 
-            $table->foreign('status_id')->references('id')->on('status')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
-        });
+                $table->foreign('status_id')->references('id')->on('status')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
+            });
+        }else{
+            Schema::create('appointments', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('patient_id')->unsigned();
+                $table->integer('office_id')->unsigned();
+                $table->dateTime('appiontment_date');
+                $table->decimal('amount',10,2);
+                $table->integer('status_id')->unsigned();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
