+++ database/migrations/2015_07_30_081819_create_workdays_table.php
@@ -12,23 +12,34 @@ class CreateWorkdaysTable extends Migration
      */
     public function up()
     {
-        Schema::create('workdays', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('office_id')->unsigned();
-            $table->tinyInteger('day_id')->unsigned();
-            $table->dateTime('start_time');
-            $table->dateTime('end_time');
-            $table->timestamps();
+        if(Schema::hasTable('offices') && Schema::hasTable('days') && Schema::hasTable('workdays')){
+            Schema::table('workdays', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('office_id')->unsigned();
+                $table->tinyInteger('day_id')->unsigned();
+                $table->dateTime('start_time');
+                $table->dateTime('end_time');
+                $table->timestamps();
 
 
-            $table->foreign('office_id')->references('id')->on('offices')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
+                $table->foreign('office_id')->references('id')->on('offices')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
 
-            $table->foreign('day_id')->references('id')->on('days')
-                    ->onUpdate('CASCADE')
-                    ->onDelete('NO ACTION');
-        });
+                $table->foreign('day_id')->references('id')->on('days')
+                        ->onUpdate('CASCADE')
+                        ->onDelete('NO ACTION');
+            });
+        }else{
+            Schema::create('workdays', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('office_id')->unsigned();
+                $table->tinyInteger('day_id')->unsigned();
+                $table->dateTime('start_time');
+                $table->dateTime('end_time');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
